GC smart-pointer optimizations
GC struct now ALSO offers reference-counted pointers that are cycle-UNsafe, which are faster than cycle-safe pointers.

Reasoning to not use std::shared_ptr can be found in "garbage_collector.hpp".

This allows for more efficient pointers when working with strings, as well as the "recursive-depth" counter used in struct function_object.